refactor: Extract CarForm from Cars component
