+++ src/components/cars/Cars.js
@@ -1,39 +1,15 @@
 import React, { useState } from "react";
 import axios from "axios";
-import {
-  Tabs,
-  Tab,
-} from "@mui/material";
+import { Tabs, Tab } from "@mui/material";
+import CarForm from "./CarForm";
 
 export default function CarInfo() {
-  const [year, setYear] = useState("");
-  const [make, setMake] = useState("");
-  const [model, setModel] = useState("");
-  const [models, setModels] = useState([]);
   const [data, setData] = useState({ recalls: [], ratings: [] });
   const [errorMessage, setErrorMessage] = useState("");
   const [hasFetchedData, setHasFetchedData] = useState(false);
   const [activeRecallTab, setActiveRecallTab] = useState(0);
-  const years = [2017, 2018, 2019, 2020, 2021, 2022];
-  const makes = ["Honda", "BMW", "Toyota", "Ford", "Audi"];
 
-  const makeModelMapping = {
-    Honda: ["Civic", "Accord", "Pilot"],
-    BMW: ["X5", "X3", "M3"],
-    Toyota: ["Corolla", "Camry", "RAV4"],
-    Ford: ["F-150", "Taurus", "Mustang"],
-    Audi: ["A5", "A6", "RS5"],
-  };
-
-  const handleMakeChange = (event) => {
-    const selectedMake = event.target.value;
-    setMake(selectedMake);
-    setModel("");
-    setModels(makeModelMapping[selectedMake]);
-    setActiveRecallTab(0);
-  };
-
-  const fetchData = async () => {
+  const fetchData = async (year, make, model) => {
     try {
       const response = await axios.get(
         `/api/fetchData?year=${year}&make=${make}&model=${model}`
@@ -101,56 +77,7 @@ export default function CarInfo() {
           Get Vehicle Information
         </h2>
 
-        <select
-          className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-          value={year}
-          onChange={(e) => setYear(e.target.value)}
-        >
-          <option value="" disabled>
-            Select Year
-          </option>
-          {years.map((y) => (
-            <option key={y} value={y}>
-              {y}
-            </option>
-          ))}
-        </select>
-        <select
-          className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-          value={make}
-          onChange={handleMakeChange}
-        >
-          <option value="" disabled>
-            Select Make
-          </option>
-          {makes.map((m) => (
-            <option key={m} value={m}>
-              {m}
-            </option>
-          ))}
-        </select>
-        <select
-          className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-          value={model}
-          onChange={(e) => setModel(e.target.value)}
-          disabled={!models.length}
-        >
-          <option value="" disabled>
-            Select Model
-          </option>
-          {models.map((m) => (
-            <option key={m} value={m}>
-              {m}
-            </option>
-          ))}
-        </select>
-
-        <button
-          onClick={fetchData}
-          className="bg-[#832C31] hover:bg-[#42191b] text-white font-bold py-2 px-4 rounded mt-5 block w-full"
-        >
-          Fetch Data
-        </button>
+        <CarForm fetchData={fetchData} />
 
         {errorMessage && (
           <p className="text-[#832C31] text-center mt-5">{errorMessage}</p>
